Quick patch for question bank re-import issue.
this function is called from the UI when saving a question, and during the question-bank import process.

the data structure of the responsetemplate is different between these
two workflows.

the correct data structure is an assoc array, which is what the UI
provides on save. the import process gives us a string here however,
which was unexpected, resulting in an "Cannot access offset of type
string on string" type error.

adding a type check here and dealing with the variation this way is a
stop-gap solution until this can be corrected on the export-side of
things.